web-animations/animation-group/animation-group-children-start-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationGroup child nodes start time test</title>
<meta name="assert" content="The start time of a child animation node of an animation group is zero">
<link rel="help" href="http://w3c.github.io/web-animations/#the-start-time-of-children-of-an-animation-group">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1, node2]);
    var player = document.timeline.play(animationGroup);

    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationGroup constructor sets start time to zero for each child');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationGroup([]);
    var node3 = new AnimationSequence([]);
    var animationGroup = new AnimationGroup([node1, node2, node3]);
    var player = document.timeline.play(animationGroup);

    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
    assert_equals(node3.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
}, 'Test AnimationGroup constructor sets start time to zero for each child. ' +
    'Children are of different types');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = newAnimation(createDiv(this));
    var node5 = new AnimationGroup([node1, node2]);
    var node6 = new AnimationSequence([node3]);
    var animationGroup = new AnimationGroup([node4, node5, node6]);
    var player = document.timeline.play(animationGroup);

    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
    assert_equals(node4.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
    assert_equals(node5.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
    assert_equals(node6.computedTiming.startTime, 0, 'The start time of a child ' +
        'animation node of an animation group should be zero');
}, 'Test AnimationGroup constructor sets start time to zero for each child. ' +
    'Test tree structure');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    animationGroup.prepend(node2);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationGroup.prepend() sets start time to zero on prepended node');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([]);
    var node4 = new AnimationSequence([]);
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    animationGroup.prepend(node2, node3, node4);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node3.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node4.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationGroup.prepend() sets start time to zero on different prepended nodes');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    animationGroup.append(node2);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationGroup.append() sets start time to zero on appended node');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([]);
    var node4 = new AnimationSequence([]);
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    animationGroup.append(node2, node3, node4);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node3.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node4.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationGroup.append() sets start time to zero on different appended nodes');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    node1.before(node2);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationNode.before() sets start time to sero on inserted node, ' +
    'if adding the node to animation group');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([]);
    var node4 = new AnimationSequence([]);
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    node1.before(node2, node3, node4);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node3.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node4.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationNode.before() sets start time to sero on inserted nodes, ' +
    'if adding nodes to animation group');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    node1.after(node2);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationNode.after() sets start time to sero on inserted node, ' +
    'if adding the node to animation group');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([]);
    var node4 = new AnimationSequence([]);
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    node1.after(node2, node3, node4);
    assert_equals(node1.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node3.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node4.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationNode.after() sets start time to sero on inserted nodes, ' +
    'if adding nodes to animation group');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    node1.replace(node2);
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationNode.replace() sets start time to sero on inserted node, ' +
    'if parent node is animation group');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([]);
    var node4 = new AnimationSequence([]);
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(animationGroup);

    node1.replace(node2, node3, node4);
    assert_equals(node2.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node3.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
    assert_equals(node4.computedTiming.startTime, 0, 'The start time of a ' +
        'child animation node of an animation group should be zero');
}, 'Test AnimationNode.replace() sets start time to sero on inserted nodes, ' +
    'if parent node is animation group');
</script>
</body>
